// Unmatched.UI.Angular/src/app/backend-api/models/ui-hero-statistics-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TitleDto } from '../models/title-dto';
import { UiPlayStyleDto } from '../models/ui-play-style-dto';
import { UiSidekickDto } from '../models/ui-sidekick-dto';
export interface UiHeroStatisticsDto {
  color?: string | null;
  deckSize?: number;
  heroId?: string;
  hp?: number;
  imageUrl?: string | null;
  isRanged?: boolean;
  kd?: number;
  lastMatchPoints?: number;
  meleeRangeImageUrl?: string | null;
  name?: string | null;
  place?: number;
  playStyle?: UiPlayStyleDto;
  points?: number;
  sidekicks?: Array<UiSidekickDto> | null;
  titles?: Array<TitleDto> | null;
  totalLooses?: number;
  totalMatches?: number;
  totalWins?: number;
}